# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WH40k Colony Manager API",
    "description": "REST API for managing Warhammer 40k Rogue Trader colonies",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/representatives": {
      "get": {
        "tags": [
          "representatives"
        ],
        "summary": "List Representatives",
        "description": "List all representatives with pagination and filtering.\n\nReturns a paginated list of representatives. Use offset/limit for pagination.",
        "operationId": "list_representatives_api_v1_representatives_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "available_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Only show unassigned representatives",
              "default": false,
              "title": "Available Only"
            },
            "description": "Only show unassigned representatives"
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RepresentativeType"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by representative type",
              "examples": [
                "judge",
                "cardinal",
                "satrap"
              ],
              "title": "Type"
            },
            "description": "Filter by representative type"
          },
          {
            "name": "name_search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search by name (case-insensitive substring match)",
              "examples": [
                "cardinal",
                "valmar"
              ],
              "title": "Name Search"
            },
            "description": "Search by name (case-insensitive substring match)"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of items to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of items to skip"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum number of items to return",
              "default": 20,
              "title": "Limit"
            },
            "description": "Maximum number of items to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_RepresentativeListItem_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "representatives"
        ],
        "summary": "Create Representative",
        "description": "Create a new representative.",
        "operationId": "create_representative_api_v1_representatives_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepresentativeCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepresentativeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/representatives/{rep_id}": {
      "get": {
        "tags": [
          "representatives"
        ],
        "summary": "Get Representative",
        "description": "Get a representative by ID.",
        "operationId": "get_representative_api_v1_representatives__rep_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "rep_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rep Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepresentativeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "representatives"
        ],
        "summary": "Update Representative",
        "description": "Update a representative (partial update).",
        "operationId": "update_representative_api_v1_representatives__rep_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "rep_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rep Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepresentativeUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepresentativeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "representatives"
        ],
        "summary": "Delete Representative",
        "description": "Delete a representative.",
        "operationId": "delete_representative_api_v1_representatives__rep_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "rep_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rep Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/representatives/{rep_id}/assign": {
      "post": {
        "tags": [
          "representatives"
        ],
        "summary": "Assign To Colony",
        "description": "Assign a representative to a colony.",
        "operationId": "assign_to_colony_api_v1_representatives__rep_id__assign_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "rep_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rep Id"
            }
          },
          {
            "name": "colony_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Colony Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepresentativeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/representatives/{rep_id}/unassign": {
      "post": {
        "tags": [
          "representatives"
        ],
        "summary": "Unassign From Colony",
        "description": "Unassign a representative from their colony.",
        "operationId": "unassign_from_colony_api_v1_representatives__rep_id__unassign_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "rep_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rep Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepresentativeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse_RepresentativeListItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RepresentativeListItem"
            },
            "type": "array",
            "title": "Items"
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "type": "object",
        "required": [
          "items",
          "meta"
        ],
        "title": "PaginatedResponse[RepresentativeListItem]"
      },
...
      "PaginationMeta": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages",
            "description": "Calculate total number of pages.",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "total",
          "offset",
          "limit",
          "has_more",
          "total_pages"
        ],
        "title": "PaginationMeta",
        "description": "Pagination metadata."
      },
...
      "RepresentativeCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "type": {
            "$ref": "#/components/schemas/RepresentativeType"
          },
          "personalities": {
            "items": {
              "$ref": "#/components/schemas/PersonalityCreate"
            },
            "type": "array",
            "title": "Personalities",
            "description": "List of 1-4 personalities. Count limit depends on 'Quite a Character' position."
          },
          "stats": {
            "$ref": "#/components/schemas/RepresentativeStatsCreate"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/Skill"
            },
            "type": "array",
            "title": "Skills"
          },
          "talents": {
            "items": {
              "$ref": "#/components/schemas/Talent"
            },
            "type": "array",
            "title": "Talents"
          }
        },
        "type": "object",
        "required": [
          "name",
          "type",
          "stats"
        ],
        "title": "RepresentativeCreate",
        "description": "Schema for creating a new representative.\n\nPersonality count rules (per Rogue Trader Table 3-6):\n- Base limit: 2 personalities maximum\n- If 'Quite a Character' is first (index 0): limit increases to 4\n- If 'Quite a Character' is second (index 1): limit increases to 3\n- Minimum: 1 personality required"
      },
      "RepresentativeListItem": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "$ref": "#/components/schemas/RepresentativeType"
          },
          "leadership_modifier": {
            "type": "integer",
            "title": "Leadership Modifier"
          },
          "assigned_to_colony_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To Colony Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "type",
          "leadership_modifier",
          "assigned_to_colony_id"
        ],
        "title": "RepresentativeListItem",
        "description": "Summary information for representative list."
      },
      "RepresentativeResponse": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "$ref": "#/components/schemas/RepresentativeType"
          },
          "personalities": {
            "items": {
              "$ref": "#/components/schemas/Personality"
            },
            "type": "array",
            "title": "Personalities"
          },
          "stats": {
            "$ref": "#/components/schemas/RepresentativeStatsCreate"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/Skill"
            },
            "type": "array",
            "title": "Skills"
          },
          "talents": {
            "items": {
              "$ref": "#/components/schemas/Talent"
            },
            "type": "array",
            "title": "Talents"
          },
          "leadership_modifier": {
            "type": "integer",
            "title": "Leadership Modifier"
          },
          "assigned_to_colony_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To Colony Id"
          },
          "assignment_change": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssignmentChangeInfo"
              },
              {
                "type": "null"
              }
            ],
            "description": "Change tracking information for assignment/unassign operations (None for other endpoints)"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "type",
          "personalities",
          "stats",
          "skills",
          "talents",
          "leadership_modifier",
          "assigned_to_colony_id"
        ],
        "title": "RepresentativeResponse",
        "description": "Full representative response."
      },
...
      "RepresentativeType": {
        "type": "string",
        "enum": [
          "satrap",
          "judge",
          "cardinal",
          "colonist_representative",
          "military_commander",
          "dynasty_member"
        ],
        "title": "RepresentativeType"
      },
      "RepresentativeUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "personalities": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PersonalityCreate"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Personalities"
          },
          "stats": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RepresentativeStatsCreate"
              },
              {
                "type": "null"
              }
            ]
          },
          "skills": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Skill"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skills"
          },
          "talents": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Talent"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Talents"
          }
        },
        "type": "object",
        "title": "RepresentativeUpdate",
        "description": "Schema for updating a representative (partial update)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token (do not include 'Bearer' prefix in the value)"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}